work-attendance-publish-service: remove database related dependencies, add kafka dep
