+++ work-attendance-publish-service/src/main/java/com/gabrielluciano/workattendancepublishservice/WorkAttendancePublishServiceApplication.java
@@ -1,0 +1,13 @@
+package com.gabrielluciano.workattendancepublishservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WorkAttendancePublishServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(WorkAttendancePublishServiceApplication.class, args);
+	}
+
+}
