+++ work-attendance-publish-service/src/main/java/com/gabrielluciano/workattendancepublishservice/application/resource/exception/StandardError.java
@@ -1,0 +1,17 @@
+package com.gabrielluciano.workattendancepublishservice.application.resource.exception;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+@Builder
+public class StandardError {
+
+    private String error;
+    private String message;
+    private String path;
+    private LocalDateTime timestamp;
+    private int status;
+}
