+++ work-attendance-publish-service/src/main/java/com/gabrielluciano/workattendancepublishservice/domain/exception/EntityNotFoundException.java
@@ -1,0 +1,14 @@
+package com.gabrielluciano.workattendancepublishservice.domain.exception;
+
+import lombok.Getter;
+
+@Getter
+public class EntityNotFoundException extends RuntimeException {
+
+    private final transient Object id;
+
+    public EntityNotFoundException(Object id, Class<?> clazz) {
+        super(String.format("Entity of type '%s' with '%s' not found", clazz.getSimpleName(), id));
+        this.id = id;
+    }
+}
