+++ work-attendance-publish-service/src/main/java/com/gabrielluciano/workattendancepublishservice/domain/service/WorkAttendanceService.java
@@ -1,0 +1,8 @@
+package com.gabrielluciano.workattendancepublishservice.domain.service;
+
+import com.gabrielluciano.workattendancepublishservice.domain.dto.CreateWorkAttendanceRequest;
+
+public interface WorkAttendanceService {
+
+    void save(CreateWorkAttendanceRequest createWorkAttendanceRequest);
+}
