+++ work-attendance-publish-service/src/test/java/com/gabrielluciano/workattendancepublishservice/WorkAttendancePublishServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.gabrielluciano.workattendancepublishservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class WorkAttendancePublishServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
