+++ work-attendance-publish-service/src/test/java/com/gabrielluciano/workattendancepublishservice/util/JsonUtils.java
@@ -1,0 +1,10 @@
+package com.gabrielluciano.workattendancepublishservice.util;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class JsonUtils {
+
+    public static String asJsonString(final Object object) throws Exception {
+        return new ObjectMapper().writeValueAsString(object);
+    }
+}
