--- work-attendance-service/src/main/java/com/gabrielluciano/workattendanceservice/WorkAttendanceServiceApplication.java
@@ -1,13 +1,0 @@
-package com.gabrielluciano.workattendanceservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WorkAttendanceServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(WorkAttendanceServiceApplication.class, args);
-	}
-
-}
