--- work-attendance-service/src/main/java/com/gabrielluciano/workattendanceservice/application/resource/exception/StandardError.java
@@ -1,17 +1,0 @@
-package com.gabrielluciano.workattendanceservice.application.resource.exception;
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-@Builder
-public class StandardError {
-
-    private String error;
-    private String message;
-    private String path;
-    private LocalDateTime timestamp;
-    private int status;
-}
