--- work-attendance-service/src/main/java/com/gabrielluciano/workattendanceservice/domain/exception/DuplicatedEntityException.java
@@ -1,14 +1,0 @@
-package com.gabrielluciano.workattendanceservice.domain.exception;
-
-import lombok.Getter;
-
-@Getter
-public class DuplicatedEntityException extends RuntimeException {
-
-    private final transient Object id;
-
-    public DuplicatedEntityException(Object id, Class<?> clazz) {
-        super(String.format("Entity of type '%s' with '%s' already exists", clazz.getSimpleName(), id));
-        this.id = id;
-    }
-}
