--- work-attendance-service/src/main/java/com/gabrielluciano/workattendanceservice/domain/service/WorkAttendanceService.java
@@ -1,11 +1,0 @@
-package com.gabrielluciano.workattendanceservice.domain.service;
-
-import com.gabrielluciano.workattendanceservice.domain.dto.CreateWorkAttendanceRequest;
-import com.gabrielluciano.workattendanceservice.domain.dto.WorkAttendanceResponse;
-
-public interface WorkAttendanceService {
-
-    void save(CreateWorkAttendanceRequest createWorkAttendanceRequest);
-
-    WorkAttendanceResponse findByCpfYearAndMonth(String cpf, Integer year, Integer month);
-}
