--- work-attendance-service/src/main/resources/application.yml
@@ -1,31 +1,0 @@
-spring:
-  application:
-    name: work-attendance-service
-  datasource:
-    url: jdbc:postgresql://localhost:5432/work_attendance_db
-    username: service
-    password: password
-  jpa:
-    hibernate:
-      ddl-auto: update
-    show-sql: true
-
-server:
-  port: 0 # random port
-
-management:
-  endpoints:
-    web:
-      exposure:
-        include: "health,info,metrics,env,caches,threaddump,logfile"
-
-eureka:
-  instance:
-    instanceId: ${spring.application.name}:${spring.application.instance_id:${random.value}}
-  client:
-    serviceUrl:
-      defaultZone: http://localhost:8761/eureka/
-
-#logging:
-#  file:
-#    name: ./logs/${spring.application.name}-log.log
