--- work-attendance-service/src/test/java/com/gabrielluciano/workattendanceservice/util/JsonUtils.java
@@ -1,10 +1,0 @@
-package com.gabrielluciano.workattendanceservice.util;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public class JsonUtils {
-
-    public static String asJsonString(final Object object) throws Exception {
-        return new ObjectMapper().writeValueAsString(object);
-    }
-}
